Code refactor (#29)
* Moved JS and CSS files to a component folder

* Fonts moved and renaming of global css file

* Favicon updated

* Changed folder names to kebab-case
