+++ public/assets/style.css
@@ -1,0 +1,14 @@
+@font-face {
+  font-family: Lato-Regular;
+  src: url(./Lato-Regular.ttf);
+}
+
+@font-face {
+  font-family: Lato-Light;
+  src: url(./Lato-Light.ttf);
+}
+
+@font-face {
+  font-family: Lato-Italic;
+  src: url(./Lato-Italic.ttf);
+}
