+++ src/App.js
@@ -1,11 +1,12 @@
 import React from "react";
 import { useSelector } from "react-redux";
-import { Headline } from "./components/Headline";
-import Chart from "./components/chart";
-import LatestNews from "./components/LatestNews";
-import KeyStats from "./components/KeyStats";
-import { Company } from "./components/Company";
-import { Footer } from "./components/Footer";
+import { Headline } from "./components/headline";
+import { Chart } from "./components/chart";
+import { LatestNews } from "./components/latest-news";
+import { KeyStats } from "./components/key-stats";
+import { Company } from "./components/company";
+import { Footer } from "./components/footer";
+import "./App.css";
 
 const App = () => {
   const selectedStock = useSelector(state => state.stockData.selectedStock);
