+++ src/components/chart/index.js
@@ -1,3 +1,1 @@
-import Chart from "./Chart";
-
-export default Chart;
+export { Chart } from "./Chart";
