+++ src/components/company/Company.css
