+++ src/components/footer/Footer.js
@@ -1,4 +1,5 @@
 import React from "react";
+import "./Footer.css";
 
 export const Footer = () => {
   return (
