+++ src/components/headline/adaptive-logo.png
