+++ src/components/headline/index.js
@@ -1,0 +1,1 @@
+export { Headline } from "./Headline";
