+++ src/components/key-stats/index.js
@@ -1,0 +1,1 @@
+export { KeyStats } from "./KeyStats";
