+++ src/components/loading/index.js
@@ -1,0 +1,1 @@
+export { Loading } from "./Loading";
