+++ src/components/market-status/index.js
@@ -1,0 +1,1 @@
+export { MarketStatus } from "./MarketStatus";
