+++ src/components/stock-ticker/StockTicker.css
