+++ src/components/stock-ticker/index.js
@@ -1,0 +1,1 @@
+export { StockTicker } from "./StockTicker";
