+++ src/index.js
@@ -1,6 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import "./styles/index.js";
+import "./styles/Global.css";
+import "./styles/Variables.css";
 import App from "./App";
 import * as serviceWorker from "./serviceWorker";
