+++ src/styles/Global.css
@@ -1,18 +1,3 @@
-@font-face {
-  font-family: Lato-Regular;
-  src: url(../assets/font/Lato-Regular.ttf);
-}
-
-@font-face {
-  font-family: Lato-Light;
-  src: url(../assets/font/Lato-Light.ttf);
-}
-
-@font-face {
-  font-family: Lato-Italic;
-  src: url(../assets/font/Lato-Italic.ttf);
-}
-
 html {
   min-height: 100%;
 }
@@ -30,18 +15,6 @@ h1.title {
   text-transform: uppercase;
   font-size: 18px;
   padding: 0 0 3px;
-}
-
-.loading-spinner {
-  border: 10px solid #ffffff31;
-  border-radius: 50%;
-  border-top: 10px solid #95a4d1;
-  width: 50px;
-  height: 50px;
-  -webkit-animation: spin 2s linear infinite;
-  animation: spin 2s linear infinite;
-  margin-left: auto;
-  margin-right: auto;
 }
 
 @keyframes spin {
