--- src/styles/index.js
@@ -1,12 +1,0 @@
-import "./Variables.css";
-import "./Global.css";
-import "./Grid.css";
-import "./Search.css";
-import "./KeyStats.css";
-import "./Headline.css";
-import "./Peers.css";
-import "./Company.css";
-import "./MarketStatus.css";
-import "./LatestNews.css";
-import "./StockTicker.css";
-import "./Footer.css";
